+++ lib/new_relic/agent/configuration/default_source.rb
@@ -211,6 +211,22 @@ module NewRelic
           end
         end
 
+        def self.convert_to_hash(value)
+          return value if value.is_a?(Hash)
+
+          if value.is_a?(String)
+            return value.split(',').each_with_object({}) do |item, hash|
+              key, value = item.split('=')
+              hash[key] = value
+            end
+          end
+
+          raise ArgumentError.new(
+            "Config value '#{value}' of " \
+            "class #{value.class} couldn't be turned into a Hash."
+          )
+        end
+
         SEMICOLON = ';'.freeze
         def self.convert_to_list_on_semicolon(value)
           case value
@@ -427,7 +443,7 @@ module NewRelic
             'Known use-case for this option is where Sinatra is running as an embedded service within another framework ' \
             'and the agent is detecting the Sinatra app and skipping the `at_exit` handler as a result. Sinatra classically ' \
             'runs the entire application in an `at_exit` block and would otherwise misbehave if the Agent\'s `at_exit` handler ' \
-            'was also installed in those circumstances.  Note: `send_data_on_exit` should also be set to `true` in  tandem with this setting.'
+            'was also installed in those circumstances. Note: `send_data_on_exit` should also be set to `true` in  tandem with this setting.'
         },
         :high_security => {
           :default => false,
@@ -588,8 +604,7 @@ module NewRelic
 
   - If you do not want the agent to capture query information, set this to `none`.
   - If you want the agent to capture all query information in its original form, set this to `raw`.
-  - When you enable [high security mode](/docs/agents/manage-apm-agents/configuration/high-security-mode), this is automatically set to `obfuscated`.
-  '
+  - When you enable [high security mode](/docs/agents/manage-apm-agents/configuration/high-security-mode), this is automatically set to `obfuscated`.'
         },
 
         :'transaction_tracer.stack_trace_threshold' => {
@@ -748,6 +763,37 @@ module NewRelic
           :type => Boolean,
           :allowed_from_server => false,
           :description => 'If `true`, the agent captures log records emitted by your application.'
+        },
+        :'application_logging.forwarding.log_level' => {
+          :default => 'debug',
+          :public => true,
+          :type => String,
+          :allowed_from_server => false,
+          :description => <<~DESCRIPTION
+            Sets the minimum level a log event must have to be forwarded to New Relic.
+
+            This is based on the integer values of Ruby's `Logger::Severity` constants: https://github.com/ruby/ruby/blob/master/lib/logger/severity.rb
+
+            The intention is to forward logs with the level given to the configuration, as well as any logs with a higher level of severity.
+
+            For example, setting this value to "debug" will forward all log events to New Relic. Setting this value to "error" will only forward log events with the levels "error", "fatal", and "unknown".
+
+            Valid values (ordered lowest to highest):
+              * "debug"
+              * "info"
+              * "warn"
+              * "error"
+              * "fatal"
+              * "unknown"
+          DESCRIPTION
+        },
+        :'application_logging.forwarding.custom_attributes' => {
+          :default => {},
+          :public => true,
+          :type => Hash,
+          :transform => DefaultSource.method(:convert_to_hash),
+          :allowed_from_server => false,
+          :description => 'A hash with key/value pairs to add as custom attributes to all log events forwarded to New Relic. If sending using an environment variable, the value must be formatted like: "key1=value1,key2=value2"'
         },
         :'application_logging.forwarding.max_samples_stored' => {
           :default => 10000,
@@ -951,7 +997,7 @@ module NewRelic
         },
         :'audit_log.path' => {
           :default => DefaultSource.audit_log_path,
-          :documentation_default => 'config/newrelic_audit.log',
+          :documentation_default => 'log/newrelic_audit.log',
           :public => true,
           :type => String,
           :allowed_from_server => false,
@@ -1320,7 +1366,7 @@ module NewRelic
           :type => String,
           :dynamic_name => true,
           :allowed_from_server => false,
-          :description => 'Controls auto-instrumentation of Curb at start up.  May be one of: `auto`, `prepend`, `chain`, `disabled`.'
+          :description => 'Controls auto-instrumentation of Curb at start up. May be one of: `auto`, `prepend`, `chain`, `disabled`.'
         },
         :'instrumentation.delayed_job' => {
           :default => 'auto',
@@ -1329,7 +1375,7 @@ module NewRelic
           :type => String,
           :dynamic_name => true,
           :allowed_from_server => false,
-          :description => 'Controls auto-instrumentation of Delayed Job at start up.  May be one of: `auto`, `prepend`, `chain`, `disabled`.'
+          :description => 'Controls auto-instrumentation of Delayed Job at start up. May be one of: `auto`, `prepend`, `chain`, `disabled`.'
         },
         :'instrumentation.elasticsearch' => {
           :default => 'auto',
@@ -1346,7 +1392,7 @@ module NewRelic
           :type => String,
           :dynamic_name => true,
           :allowed_from_server => false,
-          :description => 'Controls auto-instrumentation of Excon at start up.  May be one of: `enabled`, `disabled`.'
+          :description => 'Controls auto-instrumentation of Excon at start up. May be one of: `enabled`, `disabled`.'
         },
         :'instrumentation.grape' => {
           :default => 'auto',
@@ -1354,7 +1400,7 @@ module NewRelic
           :type => String,
           :dynamic_name => true,
           :allowed_from_server => false,
-          :description => 'Controls auto-instrumentation of Grape at start up.  May be one of: `auto`, `prepend`, `chain`, `disabled`.'
+          :description => 'Controls auto-instrumentation of Grape at start up. May be one of: `auto`, `prepend`, `chain`, `disabled`.'
         },
         :'instrumentation.grpc_client' => {
           :default => 'auto',
@@ -1363,7 +1409,7 @@ module NewRelic
           :type => String,
           :dynamic_name => true,
           :allowed_from_server => false,
-          :description => 'Controls auto-instrumentation of gRPC clients at start up.  May be one of: `auto`, `prepend`, `chain`, `disabled`.'
+          :description => 'Controls auto-instrumentation of gRPC clients at start up. May be one of: `auto`, `prepend`, `chain`, `disabled`.'
         },
         :'instrumentation.grpc.host_denylist' => {
           :default => [],
@@ -1380,7 +1426,7 @@ module NewRelic
           :type => String,
           :dynamic_name => true,
           :allowed_from_server => false,
-          :description => 'Controls auto-instrumentation of gRPC servers at start up.  May be one of: `auto`, `prepend`, `chain`, `disabled`.'
+          :description => 'Controls auto-instrumentation of gRPC servers at start up. May be one of: `auto`, `prepend`, `chain`, `disabled`.'
         },
         :'instrumentation.httpclient' => {
           :default => 'auto',
@@ -1389,7 +1435,7 @@ module NewRelic
           :type => String,
           :dynamic_name => true,
           :allowed_from_server => false,
-          :description => 'Controls auto-instrumentation of HTTPClient at start up.  May be one of: `auto`, `prepend`, `chain`, `disabled`.'
+          :description => 'Controls auto-instrumentation of HTTPClient at start up. May be one of: `auto`, `prepend`, `chain`, `disabled`.'
         },
         :'instrumentation.httprb' => {
           :default => 'auto',
@@ -1398,7 +1444,7 @@ module NewRelic
           :type => String,
           :dynamic_name => true,
           :allowed_from_server => false,
-          :description => 'Controls auto-instrumentation of http.rb gem at start up.  May be one of: `auto`, `prepend`, `chain`, `disabled`.'
+          :description => 'Controls auto-instrumentation of http.rb gem at start up. May be one of: `auto`, `prepend`, `chain`, `disabled`.'
         },
         :'instrumentation.logger' => {
           :default => instrumentation_value_from_boolean(:'application_logging.enabled'),
@@ -1407,7 +1453,7 @@ module NewRelic
           :type => String,
           :dynamic_name => true,
           :allowed_from_server => false,
-          :description => 'Controls auto-instrumentation of Ruby standard library Logger at start up.  May be one of: `auto`, `prepend`, `chain`, `disabled`.'
+          :description => 'Controls auto-instrumentation of Ruby standard library Logger at start up. May be one of: `auto`, `prepend`, `chain`, `disabled`.'
         },
         :'instrumentation.memcache' => {
           :default => 'auto',
@@ -1415,7 +1461,7 @@ module NewRelic
           :type => String,
           :dynamic_name => true,
           :allowed_from_server => false,
-          :description => 'Controls auto-instrumentation of dalli gem for Memcache at start up.  May be one of: `auto`, `prepend`, `chain`, `disabled`.'
+          :description => 'Controls auto-instrumentation of dalli gem for Memcache at start up. May be one of: `auto`, `prepend`, `chain`, `disabled`.'
         },
         :'instrumentation.memcached' => {
           :default => 'auto',
@@ -1424,7 +1470,7 @@ module NewRelic
           :type => String,
           :dynamic_name => true,
           :allowed_from_server => false,
-          :description => 'Controls auto-instrumentation of memcached gem for Memcache at start up.  May be one of: `auto`, `prepend`, `chain`, `disabled`.'
+          :description => 'Controls auto-instrumentation of memcached gem for Memcache at start up. May be one of: `auto`, `prepend`, `chain`, `disabled`.'
         },
         :'instrumentation.memcache_client' => {
           :default => 'auto',
@@ -1433,7 +1479,7 @@ module NewRelic
           :type => String,
           :dynamic_name => true,
           :allowed_from_server => false,
-          :description => 'Controls auto-instrumentation of memcache-client gem for Memcache at start up.  May be one of: `auto`, `prepend`, `chain`, `disabled`.'
+          :description => 'Controls auto-instrumentation of memcache-client gem for Memcache at start up. May be one of: `auto`, `prepend`, `chain`, `disabled`.'
         },
         :'instrumentation.mongo' => {
           :default => 'enabled',
@@ -1442,7 +1488,7 @@ module NewRelic
           :type => String,
           :dynamic_name => true,
           :allowed_from_server => false,
-          :description => 'Controls auto-instrumentation of Mongo at start up.  May be one of: `enabled`, `disabled`.'
+          :description => 'Controls auto-instrumentation of Mongo at start up. May be one of: `enabled`, `disabled`.'
         },
         :'instrumentation.net_http' => {
           :default => 'auto',
@@ -1451,7 +1497,7 @@ module NewRelic
           :type => String,
           :dynamic_name => true,
           :allowed_from_server => false,
-          :description => 'Controls auto-instrumentation of `Net::HTTP` at start up.  May be one of: `auto`, `prepend`, `chain`, `disabled`.'
+          :description => 'Controls auto-instrumentation of `Net::HTTP` at start up. May be one of: `auto`, `prepend`, `chain`, `disabled`.'
         },
         :'instrumentation.puma_rack' => {
           :default => value_of(:'instrumentation.rack'),
@@ -1462,7 +1508,7 @@ module NewRelic
           :allowed_from_server => false,
           :description => 'Controls auto-instrumentation of `Puma::Rack`. When enabled, the agent hooks into the ' \
                            '`to_app` method in `Puma::Rack::Builder` to find gems to instrument during ' \
-                           'application startup.  May be one of: `auto`, `prepend`, `chain`, `disabled`.'
+                           'application startup. May be one of: `auto`, `prepend`, `chain`, `disabled`.'
         },
         :'instrumentation.puma_rack_urlmap' => {
           :default => value_of(:'instrumentation.rack_urlmap'),
@@ -1471,7 +1517,7 @@ module NewRelic
           :type => String,
           :dynamic_name => true,
           :allowed_from_server => false,
-          :description => 'Controls auto-instrumentation of `Puma::Rack::URLMap` at start up.  May be one of: `auto`, `prepend`, `chain`, `disabled`.'
+          :description => 'Controls auto-instrumentation of `Puma::Rack::URLMap` at start up. May be one of: `auto`, `prepend`, `chain`, `disabled`.'
         },
         :'instrumentation.rack' => {
           :default => 'auto',
@@ -1482,7 +1528,7 @@ module NewRelic
           :allowed_from_server => false,
           :description => 'Controls auto-instrumentation of Rack. When enabled, the agent hooks into the ' \
                            '`to_app` method in `Rack::Builder` to find gems to instrument during ' \
-                           'application startup.  May be one of: `auto`, `prepend`, `chain`, `disabled`.'
+                           'application startup. May be one of: `auto`, `prepend`, `chain`, `disabled`.'
         },
         :'instrumentation.rack_urlmap' => {
           :default => 'auto',
@@ -1491,7 +1537,7 @@ module NewRelic
           :type => String,
           :dynamic_name => true,
           :allowed_from_server => false,
-          :description => 'Controls auto-instrumentation of `Rack::URLMap` at start up.  May be one of: `auto`, `prepend`, `chain`, `disabled`.'
+          :description => 'Controls auto-instrumentation of `Rack::URLMap` at start up. May be one of: `auto`, `prepend`, `chain`, `disabled`.'
         },
         :'instrumentation.rake' => {
           :default => 'auto',
@@ -1499,7 +1545,7 @@ module NewRelic
           :type => String,
           :dynamic_name => true,
           :allowed_from_server => false,
-          :description => 'Controls auto-instrumentation of rake at start up.  May be one of: `auto`, `prepend`, `chain`, `disabled`.'
+          :description => 'Controls auto-instrumentation of rake at start up. May be one of: `auto`, `prepend`, `chain`, `disabled`.'
         },
         :'instrumentation.redis' => {
           :default => 'auto',
@@ -1507,7 +1553,7 @@ module NewRelic
           :type => String,
           :dynamic_name => true,
           :allowed_from_server => false,
-          :description => 'Controls auto-instrumentation of Redis at start up.  May be one of: `auto`, `prepend`, `chain`, `disabled`.'
+          :description => 'Controls auto-instrumentation of Redis at start up. May be one of: `auto`, `prepend`, `chain`, `disabled`.'
         },
         :'instrumentation.resque' => {
           :default => 'auto',
@@ -1516,7 +1562,7 @@ module NewRelic
           :type => String,
           :dynamic_name => true,
           :allowed_from_server => false,
-          :description => 'Controls auto-instrumentation of resque at start up.  May be one of: `auto`, `prepend`, `chain`, `disabled`.'
+          :description => 'Controls auto-instrumentation of resque at start up. May be one of: `auto`, `prepend`, `chain`, `disabled`.'
         },
         :'instrumentation.sinatra' => {
           :default => 'auto',
@@ -1524,7 +1570,7 @@ module NewRelic
           :type => String,
           :dynamic_name => true,
           :allowed_from_server => false,
-          :description => 'Controls auto-instrumentation of Sinatra at start up.  May be one of: `auto`, `prepend`, `chain`, `disabled`.'
+          :description => 'Controls auto-instrumentation of Sinatra at start up. May be one of: `auto`, `prepend`, `chain`, `disabled`.'
         },
         :'instrumentation.thread' => {
           :default => 'auto',
@@ -1563,7 +1609,7 @@ module NewRelic
           :type => String,
           :dynamic_name => true,
           :allowed_from_server => false,
-          :description => 'Controls auto-instrumentation of Typhoeus at start up.  May be one of: `auto`, `prepend`, `chain`, `disabled`.'
+          :description => 'Controls auto-instrumentation of Typhoeus at start up. May be one of: `auto`, `prepend`, `chain`, `disabled`.'
         },
         # Message tracer
         :'message_tracer.segment_parameters.enabled' => {
@@ -1622,7 +1668,7 @@ module NewRelic
           'This configuration option converts the Array to a RegEx list. If you\'d like ' \
           'to allow all tasks by default, use `rake.tasks: [.+]`. No rake tasks will be ' \
           'instrumented unless they\'re added to this list. For more information, ' \
-          'visit the (New Relic Rake Instrumentation docs)[/docs/apm/agents/ruby-agent/background-jobs/rake-instrumentation].'
+          'visit the [New Relic Rake Instrumentation docs](/docs/apm/agents/ruby-agent/background-jobs/rake-instrumentation).'
         },
         :'rake.connect_timeout' => {
           :default => 10,